Add files via upload
The first lab introduces the learning how to use the terminal to access the mars account. Creating directories and subdirectories for labs to be added to for file transfer and submissions to the professor.  While learning how to use Linux's command-line interface and Pico's text editor along. A simple "Hello world!" was not the meat of his lab, but developing a comprehension for a new language that was the steep learning curve of the class. I look forward to the actual meat of every lab assignment and homework!